requirements.txt
dash==2.17.1
dash_bootstrap_components==1.6.0
dash_daq==0.5.0
numpy==2.1.0
opencv_python==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
plotly==5.22.0
Requests==2.32.3
skimage==0.0
torch==2.3.0+cu118
transformers==4.43.2
